refactor: move software-update + system-logs to components/system-settings
Both components route from /system-settings/software-update/ and
/system-settings/logs/ respectively, but historically lived under
components/monitoring/ alongside the VPN/watchdog/alerts cards that
actually render under /monitoring/*. Aligns the directory layout
with the URL structure already followed by ssh-password, bandwidth-
monitor, config-backup, and at-terminal.

Internal imports use @/ absolute paths or ./sibling relative paths —
no import edits needed inside the moved files. Only the two
app/system-settings/*/page.tsx consumers update their path.

Unblocks the deferred Plan 7A i18n work (software-update + logs) —
the translated components will now live in the expected namespace.
