+++ app/system-settings/software-update/page.tsx
@@ -1,4 +1,4 @@
-import SoftwareUpdateComponent from "@/components/monitoring/software-update/software-update";
+import SoftwareUpdateComponent from "@/components/system-settings/software-update/software-update";
 
 const SoftwareUpdatePage = () => {
   return <SoftwareUpdateComponent />;
